New version under new name
Repository was transferred to github.com/intra2net and the plugin
is now called "Note Diff". The copyright notice and other strings
were fixed accordingly.

The version was adjusted such that 2.0.0 corresponds to the
version of the plugin that works on MantisBT v2 and a Github
Actions workflow was added to generate tarballs and zip packages
when tags are pushed upstream.